move applications to be deprecated
