+++ spec/deprecated/application.js
